Pre-launch audit, Pass 1: build and run verification
Clean-clone install, dev server, and every documented CI gate
(typecheck/lint/test/build, Python tests, ruff, mypy, gen:contracts:check)
run for real with pasted output, not read-and-assumed. Fixed everything
red found along the way:

- ruff: 20 real violations across 3 Python packages, including a genuine
  F821 (multiprocessing referenced in a module-level type annotation but
  only imported locally) in services/cpp-executor/instrument.py.
- mypy: discovered the documented `mypy apps/api/app` command (run from
  repo root, per CLAUDE.md) silently never applies that package's own
  strict=true config — mypy's config discovery is CWD-relative, not
  target-path-relative. Run correctly, this surfaced 82 real strict-mode
  errors across apps/api and services/executor/tracer.py (the most
  safety-critical file in the backend), plus a services/cpp-executor
  package that had never had a [tool.mypy] section at all. All fixed;
  all three packages now genuinely mypy-clean under strict mode.
- gen:contracts:check: fixed a real ENOENT (generate.mjs assumed a bare
  `uv` binary on PATH; this sandbox only has it via `pip install uv`) by
  resolving uv robustly at call time. Verified byte-identical generated
  output before/after — zero contract-shape change.
- Documented, not fixed: the executor container in docker-compose.yml and
  its Dockerfile carry none of PRD §5's mandated hardening flags
  (--network none, --pids-limit, --memory, --cap-drop ALL, non-root uid) —
  a real launch blocker, not a Pass-1-fixable config typo. docker itself
  isn't installed in this sandbox, so the compose stack could not be run
  to confirm further.

Full findings, evidence, and the one flagged packages/contracts/ edit
(generate.mjs only — a build-tooling fix, not a shape change) are in
docs/AUDIT.md.

+++ apps/api/app/rag/concept_store.py
@@ -37,7 +37,7 @@ def _cosine_similarity(a: list[float], b: list[float]) -> float:
     norm_b = sum(x * x for x in b) ** 0.5
     if not norm_a or not norm_b:
         return 0.0
-    return dot / (norm_a * norm_b)
+    return float(dot / (norm_a * norm_b))
 
 
 class InMemoryConceptStore:
